# Do patch-based building for SimpleConcat scenarios