--- a/BENCHMARKS.md
+++ b/BENCHMARKS.md
@@ -13,6 +13,7 @@ This will run benchmarks for initial builds and rebuilds. You can optionally pro
 * `FILE_COUNT` (number, default: 10) - Specifies how many files to concat together in the fixture.
 * `DEPTH` (number, default: 1) - Specifies how many sub-directories of files to place in the fixture. Each directory level will have `FILE_COUNT` files.
 * `LARGE_FILES` (boolean, default: false) - Specifies whether to use fixture files with large memory footprints (>1mb).
+* `SOURCE_MAPS` (boolean, default: true) - Specifies whether to use sourcemap concat.
 
 So, if you want to run many iterations testing the performance of concatenating several large files in a nested structure, you'd run something like: