--- a/bench/suites/rebuild.js
+++ b/bench/suites/rebuild.js
@@ -18,7 +18,10 @@ module.exports = {
     );
 
     let node = concat(this.fixture.directory, {
-      outputFile: '/result.js'
+      outputFile: '/result.js',
+      sourceMapConfig: {
+        enabled: process.env.SOURCE_MAPS !== 'false'
+      }
     });
 
     this.builder = new Builder(node);