--- a/lib/strategies/simple.js
+++ b/lib/strategies/simple.js
@@ -1,0 +1,146 @@
+'use strict';
+var fs = require('fs-extra');
+var path = require('path');
+var findIndex = require('find-index');
+var merge = require('lodash.merge');
+
+function contentMapper(entry) {
+  return entry.content;
+}
+
+function SimpleConcat(attrs) {
+  this.separator = attrs.separator || '';
+  this.inputDir = attrs.inputDir;
+  this.allowNone = attrs.allowNone;
+
+  this.header = attrs.header;
+  this.headerFiles = attrs.headerFiles || [];
+  this.footerFiles = attrs.footerFiles || [];
+  this.footer = attrs.footer;
+
+  // Internally, we represent the concatenation as a series of entries. These
+  // entries have a 'file' attribute for lookup/sorting and a 'content' property
+  // which represents the value to be used in the concatenation.
+  this._internal = [];
+
+  // We represent the header/footer files as empty at first so that we don't
+  // have to figure out order when patching
+  this._internalHeaderFiles = this.headerFiles.map(function(file) { return { file: file } });
+  this._internalFooterFiles = this.footerFiles.map(function(file) { return { file: file } });
+}
+
+SimpleConcat.isPatchBased = true;
+
+SimpleConcat.prototype = merge(SimpleConcat.prototype, {
+  /**
+   * Finds the index of the given file in the internal data structure.
+   */
+  _findIndexOf: function(file) {
+    return findIndex(this._internal, function(entry) { return entry.file === file; });
+  },
+
+  /**
+   * Updates the contents of a header file.
+   */
+  _updateHeaderFile: function(fileIndex, content) {
+    this._internalHeaderFiles[fileIndex].content = content;
+  },
+
+  /**
+   * Updates the contents of a footer file.
+   */
+  _updateFooterFile: function(fileIndex, content) {
+    this._internalFooterFiles[fileIndex].content = content;
+  },
+
+  /**
+   * Determines if the given file is a header or footer file, and if so updates
+   * it with the given contents.
+   */
+  _handleHeaderOrFooterFile: function(file, content) {
+    var headerFileIndex = this.headerFiles.indexOf(file);
+    if (headerFileIndex !== -1) {
+      this._updateHeaderFile(headerFileIndex, content);
+      return true;
+    }
+
+    var footerFileIndex = this.footerFiles.indexOf(file);
+    if (footerFileIndex !== -1) {
+      this._updateFooterFile(footerFileIndex, content);
+      return true;
+    }
+
+    return false;
+  },
+
+  addFile: function(file) {
+    var content = fs.readFileSync(path.join(this.inputDir, file), 'UTF-8');
+
+    if (this._handleHeaderOrFooterFile(file, content)) {
+      return;
+    }
+
+    var entry = {
+      file: file,
+      content: content
+    };
+
+    var index = findIndex(this._internal, function(entry) { return entry.file > file; });
+    if (index === -1) {
+      this._internal.push(entry);
+    } else {
+      this._internal.splice(index, 0, entry);
+    }
+  },
+
+  updateFile: function(file) {
+    var content = fs.readFileSync(path.join(this.inputDir, file), 'UTF-8');
+
+    if (this._handleHeaderOrFooterFile(file, content)) {
+      return;
+    }
+
+    var index = this._findIndexOf(file);
+
+    if (index === -1) {
+      throw new Error('Trying to update ' + file + ' but it has not been read before');
+    }
+
+    this._internal[index].content = content;
+  },
+
+  removeFile: function(file) {
+    if (this._handleHeaderOrFooterFile(file, '')) {
+      return;
+    }
+
+    var index = this._findIndexOf(file);
+
+    if (index === -1) {
+      throw new Error('Trying to remove ' + file + ' but it did not previously exist');
+    }
+
+    this._internal.splice(index, 1);
+  },
+
+  write: function(outputFile) {
+    // DISCUSS IN REVIEW: Should we store a concatenated string and splice it
+    // with updates? Would require keeping pointers to the start and end of
+    // files and updating those during operations.
+    var content = [].concat(
+      this.header,
+      this._internalHeaderFiles.map(contentMapper),
+      this._internal.map(contentMapper),
+      this._internalFooterFiles.map(contentMapper),
+      this.footer
+    ).filter(Boolean).join(this.separator);
+
+    if (!content && !this.allowNone) {
+      throw Error('Concatenation result is empty.');
+    }
+
+    fs.outputFileSync(outputFile, content);
+  },
+});
+
+module.exports = SimpleConcat;