--- a/lib/strategies/simple.js
+++ b/lib/strategies/simple.js
@@ -1,6 +1,4 @@
 'use strict';
-var fs = require('fs-extra');
-var path = require('path');
 var findIndex = require('find-index');
 var merge = require('lodash.merge');
 
@@ -10,8 +8,6 @@ function contentMapper(entry) {
 
 function SimpleConcat(attrs) {
   this.separator = attrs.separator || '';
-  this.inputDir = attrs.inputDir;
-  this.allowNone = attrs.allowNone;
 
   this.header = attrs.header;
   this.headerFiles = attrs.headerFiles || [];
@@ -73,9 +69,7 @@ SimpleConcat.prototype = merge(SimpleConcat.prototype, {
     return false;
   },
 
-  addFile: function(file) {
-    var content = fs.readFileSync(path.join(this.inputDir, file), 'UTF-8');
-
+  addFile: function(file, content) {
     if (this._handleHeaderOrFooterFile(file, content)) {
       return;
     }
@@ -93,9 +87,7 @@ SimpleConcat.prototype = merge(SimpleConcat.prototype, {
     }
   },
 
-  updateFile: function(file) {
-    var content = fs.readFileSync(path.join(this.inputDir, file), 'UTF-8');
-
+  updateFile: function(file, content) {
     if (this._handleHeaderOrFooterFile(file, content)) {
       return;
     }
@@ -123,7 +115,7 @@ SimpleConcat.prototype = merge(SimpleConcat.prototype, {
     this._internal.splice(index, 1);
   },
 
-  write: function(outputFile) {
+  result: function(outputFile) {
     // DISCUSS IN REVIEW: Should we store a concatenated string and splice it
     // with updates? Would require keeping pointers to the start and end of
     // files and updating those during operations.
@@ -135,11 +127,7 @@ SimpleConcat.prototype = merge(SimpleConcat.prototype, {
       this.footer
     ).filter(Boolean).join(this.separator);
 
-    if (!content && !this.allowNone) {
-      throw Error('Concatenation result is empty.');
-    }
-
-    fs.outputFileSync(outputFile, content);
+    return content;
   },
 });