--- a/lib/strategies/simple.js
+++ b/lib/strategies/simple.js
@@ -115,10 +115,18 @@ SimpleConcat.prototype = merge(SimpleConcat.prototype, {
     this._internal.splice(index, 1);
   },
 
-  result: function(outputFile) {
-    // DISCUSS IN REVIEW: Should we store a concatenated string and splice it
-    // with updates? Would require keeping pointers to the start and end of
-    // files and updating those during operations.
+  fileSizes: function() {
+    return [].concat(
+      this._internalHeaderFiles,
+      this._internal,
+      this._internalFooterFiles
+    ).reduce(function(sizes, entry) {
+      sizes[entry.file] = entry.content.length;
+      return sizes;
+    }, {});
+  },
+
+  result: function() {
     var content = [].concat(
       this.header,
       this._internalHeaderFiles.map(contentMapper),