--- a/package.json
+++ b/package.json
@@ -10,7 +10,7 @@
     "lib/"
   ],
   "scripts": {
-    "test": "mocha",
+    "test": "mocha --recursive",
     "test:debug": "mocha --no-timeouts debug",
     "bench": "node bench/info && bench bench/suites -i 10"
   },
@@ -38,6 +38,7 @@
     "broccoli-stew": "^1.3.3",
     "ensure-posix-path": "^1.0.2",
     "fast-sourcemap-concat": "^1.0.1",
+    "find-index": "^1.1.0",
     "fs-extra": "^1.0.0",
     "fs-tree-diff": "^0.5.6",
     "lodash.merge": "^4.3.0",