**Work-in-progress**

This PR changes the internal build mechanism to be patch-based. This means that rebuilds will function as a set of operations applied to the previously constructed concatenation. This should likely have negligible impact on initial builds and should make most rebuilds faster as we will only have to read from the file system for files that have changed.

This will leverage the patch format from [fs-tree-diff](https://github.com/stefanpenner/fs-tree-diff).

Tasks:
- [x] Fix failing tests
- [x] Add additional tests
- [x] Cleanup code
- [x] Document new approach
- [x] Run benchmarks (from #101)
- [x] Add concat-stats