@trentmwillis 2017-02-03
Some benchmark results:
```bash
FILE_COUNT=1000 DEPTH=3
# old approach
Average time for build over 5 iterations: 391.2ms
Average time for rebuild over 5 iterations: 311.4ms

# new approach
Average time for build over 5 iterations: 435.4ms
Average time for rebuild over 5 iterations: 274.4ms

FILE_COUNT=10 DEPTH=3
# old approach
Average time for build over 1000 iterations: 4.435ms
Average time for rebuild over 1000 iterations: 4.192ms

# new approach
Average time for build over 1000 iterations: 3.242ms
Average time for rebuild over 1000 iterations: 1.888ms
```