@trentmwillis 2017-02-11
@stefanpenner updated with your suggestions. I'll try to re-run benchmarks sometime soon and hopefully squeeze whatever remaining perf we can from this approach.