@trentmwillis 2017-02-15
Fixed an issue with the rebuild benchmarks (was invalidating way more than it should have been) and re-ran them:

```bash
SOURCE_MAPS=false FILE_COUNT=1000 DEPTH=3 yarn run bench

# new approach
Average time for build over 10 iterations: 422.6ms
Average time for rebuild over 10 iterations: 82.4ms

#old approach
Average time for build over 10 iterations: 322.3ms
Average time for rebuild over 10 iterations: 276.9ms
```

```bash
SOURCE_MAPS=false FILE_COUNT=10 DEPTH=10 yarn run bench

# new approach
Average time for build over 10 iterations: 11ms
Average time for rebuild over 10 iterations: 4.8ms

# old approach
Average time for build over 10 iterations: 15.3ms
Average time for rebuild over 10 iterations: 15.7ms
```

So rebuilds are faster across the board, but initial builds are slower when dealing with lots of files (which was kind of expected).