@stefanpenner 2017-02-15
@trentmwillis how do the first iterations compare? e.g. when the FS is not yet in memory?