@trentmwillis 2017-02-16
@stefanpenner re-ran with only the first iterations:

```bash
SOURCE_MAPS=false FILE_COUNT=1000 DEPTH=3 yarn run bench

# new approach
build: 387ms
rebuild: 85ms

# old approach
build: 300ms
rebuild: 282.5ms
```
```bash
SOURCE_MAPS=false FILE_COUNT=10 DEPTH=10 yarn run bench
# old approach
build: 22ms
rebuild: 17.5ms
```