@stefanpenner 2017-02-16
new: `rebuild: 85ms`
old: `rebuild: 282.5ms`