@trentmwillis 2017-02-16
@stefanpenner I believe this should be good to land. I restored the concat-stats functionality.