@stefanpenner 2017-02-16
released as `v3.2.0`